refactor(stt): extract cloud backends behind RemoteTranscriber
Move the WisprFlow+Kirinki and OpenAI remote logic out of
HybridTranscriptionService into RemoteTranscriber strategies selected by
cloudProvider. The router now owns only local/remote mode routing; adding a
cloud backend is a new RemoteTranscriber plus its registration, with no change
to the routing code.
