+++ composeApp/src/androidInstrumentedTest/kotlin/coredevices/coreapp/evals/RingRecordingE2ETest.kt
@@ -49,6 +49,7 @@ import coredevices.util.transcription.CactusModelPathProvider
 import coredevices.util.transcription.CactusTranscriptionService
 import coredevices.util.transcription.HybridTranscriptionService
 import coredevices.util.transcription.KirinkiTranscriptionService
+import coredevices.util.transcription.OpenAiRemoteTranscriber
 import coredevices.util.transcription.OpenAiTranscriptionService
 import coredevices.util.transcription.NoOpInferenceBoost
 import coredevices.util.transcription.TranscriptionService
@@ -63,6 +64,7 @@ import coredevices.firestore.PebbleUser
 import coredevices.libindex.database.repository.RingTransferRepository
 import coredevices.ring.agent.builtin_servlets.messaging.ApprovedBeeperContact
 import coredevices.util.transcription.WisprFlowRESTTranscriptionService
+import coredevices.util.transcription.WisprFlowRemoteTranscriber
 import dev.gitlive.firebase.Firebase
 import dev.gitlive.firebase.auth.auth
 import kotlinx.coroutines.CompletableJob
@@ -582,8 +584,14 @@ class RingRecordingE2ETest {
         single {
             CactusTranscriptionService(get(), get<CactusModelPathProvider>(), get(), NoOpInferenceBoost())
         }
+        single { WisprFlowRemoteTranscriber(get(), get(), get()) }
+        single { OpenAiRemoteTranscriber(get(), get()) }
         single {
-            HybridTranscriptionService(get(), get(), get(), get(), get(), get())
+            HybridTranscriptionService(
+                coreConfigFlow = get(),
+                cactus = get(),
+                remoteTranscribers = setOf(get<WisprFlowRemoteTranscriber>(), get<OpenAiRemoteTranscriber>()),
+            )
         } bind TranscriptionService::class
 
         // MCP tools
